refactoring GameBoardPositionInvoker
